+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - current product default value
 - customer email default value
 - hidden text input type block
-- form code parameter for widget
+- form identifier parameter for widget
 
 ## [102.0.23]
 ### Fixed